@@ -1,4 +1,4 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
 require 'rails/all'
 
@@ -8,23 +8,8 @@ Bundler.require(*Rails.groups)
 
 module Theodinproject
   class Application < Rails::Application
-    config.assets.initialize_on_precompile = false
-    config.active_support.escape_html_entities_in_json = true
-    config.assets.enabled = true
-    config.assets.version = '1.0'
-    # config.i18n.enforce_available_locales = false
-
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
-
-    # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-    # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
-    # config.time_zone = 'Central Time (US & Canada)'
-
-    # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
-    # config.i18n.default_locale = :de
-    config.active_record.raise_in_transactional_callbacks = true
   end
 end
