@@ -10,19 +10,19 @@ RSpec.describe UsersController do
 
   describe 'GET show' do
     it 'renders the users profile page' do
-      get :show, id: user.id
+      get :show, params: { id: user.id }
       expect(response).to render_template(:show)
     end
   end
 
   describe 'GET edit' do
     it 'assigns @edit' do
-      get :edit, id: user.id
+      get :edit, params: { id: user.id }
       expect(assigns(:edit)).to eql(true)
     end
 
     it 'renders the users profile' do
-      get :edit, id: user.id
+      get :edit, params: { id: user.id }
       expect(response).to render_template(:show)
     end
   end
@@ -30,23 +30,23 @@ RSpec.describe UsersController do
   describe 'POST update' do
 
     it 'displays a success message' do
-      post :update, id: user.id, user: { email: 'kevin@example.com' }
+      post :update, params: { id: user.id, user: { email: 'kevin@example.com' } }
       expect(flash[:success]).to eql('Your profile was updated successfully')
     end
 
     it 'redirects to the users profile' do
-      post :update, id: user.id, user: { email: 'kevin@example.com' }
+      post :update, params: { id: user.id, user: { email: 'kevin@example.com' } }
       expect(response).to redirect_to("/users/#{user.id}")
     end
 
     context 'when there are errors' do
       it 'displays a error message' do
-        post :update, id: user.id, user: { email: 'kevin@' }
+        post :update, params: { id: user.id, user: { email: 'kevin@' } }
         expect(flash[:error]).to eql('We could not update your profile. Errors: ["Email is invalid"]')
       end
 
       it 'renders the users profile page' do
-        post :update, id: user.id, user: { email: 'kevin@' }
+        post :update, params: { id: user.id, user: { email: 'kevin@' } }
         expect(response).to render_template(:show)
       end
     end
@@ -81,21 +81,22 @@ RSpec.describe UsersController do
     before do
       allow(controller).to receive(:current_user).and_return(user)
       allow(user).to receive(:send_confirmation_link)
+      allow(controller).to receive(:request).and_return(request)
       allow(request).to receive(:referer).and_return('/courses')
     end
 
     it 'sends the confirmation instructions' do
       expect(user).to receive(:send_confirmation_instructions)
-      get :send_confirmation_link, id: user.id
+      get :send_confirmation_link, params: { id: user.id }
     end
 
     it 'displays a notice' do
-      get :send_confirmation_link, id: user.id
+      get :send_confirmation_link, params: { id: user.id }
       expect(flash[:notice]).to eql('Confirmation instructions have been sent to your email address!')
     end
 
     it 'redirects to the referer' do
-      get :send_confirmation_link, id: user.id
+      get :send_confirmation_link, params: { id: user.id }
       expect(response).to redirect_to('/courses')
     end
   end
